wip: migrations refactoring
