+++ pallets/did/src/errors.rs
@@ -25,8 +25,6 @@ pub enum DidError {
 	StorageError(StorageError),
 	/// See [SignatureError].
 	SignatureError(SignatureError),
-	/// See [UrlError].
-	UrlError(UrlError),
 	/// See [InputError].
 	InputError(InputError),
 	/// An error that is not supposed to take place, yet it happened.
@@ -42,12 +40,6 @@ impl From<StorageError> for DidError {
 impl From<InputError> for DidError {
 	fn from(err: InputError) -> Self {
 		DidError::InputError(err)
-	}
-}
-
-impl From<UrlError> for DidError {
-	fn from(err: UrlError) -> Self {
-		DidError::UrlError(err)
 	}
 }
 
@@ -90,15 +82,6 @@ pub enum SignatureError {
 	TransactionExpired,
 }
 
-/// Error generated when validating a byte-encoded endpoint URL.
-#[derive(Debug, Eq, PartialEq)]
-pub enum UrlError {
-	/// The URL specified is not ASCII-encoded.
-	InvalidUrlEncoding,
-	/// The URL specified is not properly formatted.
-	InvalidUrlScheme,
-}
-
 /// Error generated when some extrinsic input does not respect the pallet's
 /// constraints.
 #[derive(Debug, Eq, PartialEq)]
@@ -109,8 +92,4 @@ pub enum InputError {
 	/// A number of new verification keys to remove greater than the maximum
 	/// allowed has been provided.
 	MaxVerificationKeysToRemoveLimitExceeded,
-	/// A URL longer than the maximum size allowed has been provided.
-	MaxUrlLengthExceeded,
-	/// More than the maximum number of URLs have been specified.
-	MaxUrlsCountExceeded,
 }
